factor out pages
